requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
mysql-connector-python==8.0.25
oauthlib==3.1.0
protobuf==3.17.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.5
